refactor(model): Split Thread Switcher Types
Keep the model thread switcher interface and its supporting implementations in separate files so the #640 bridge follows common .NET file organization conventions without changing behavior.
